fix(BulletChart, ComparisonLine): new file ComparisonLine, multiple fixes in BulletChart
